requirements.txt
bibtexparser==1.4.3
openai>=1.12.0
pandas>=2.2.0
tabulate>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
